0-pad signature elements
When concatenating signatures, R and S are converted to signed then
unsigned integers. In the process, if either R or S started with a
0 byte, it gets dropped. The result is that while concatSig should
always return a hex encoded 65 byte string, in some circumstances it
can end up shorter.

I apologize for not providing a test case. The only way I've found to
reproduce this involves signing a particular message with a key that
I need to keep private, and concatSig is not directly exported, so
I can't test it directly.