eth_getFilterLogs returns [] when given non log id
Prior to this change if you call eth_getFilterLogs with a filter ID that
is valid but does not match a log filter, it attempts to return all
results from the filter in question anyway. Since block filters don't
implement this functionality, this caused a crash when the ID matched
that of a block filter. This is causing problems for some ganache users,
as indicated in trufflesuite/ganache#879.