Replace scrypt with scryptsy (#425)
Installing dependencies on newer Macs does not seem to work due to a
failure in compiling the `scrypt` library. This library is indirectly
brought in via `ganache-core`. The following dependency tree looks like:

    ganache-core
      ethereumjs-wallet
        scrypt.js
          scrypt

This commit upgrades `ganache-core` from 2.10.2 to 2.13.2 and upgrades
`ethereumjs-wallet` from 0.6.3 to 0.6.5. The new version of
`ethereumjs-wallet` replaces `scrypt.js` with `scryptsy`, which does not
rely on `scrypt`, but provides a pure-JS implementation. The new
dependency tree now looks like:

    ganache-core
      ethereumjs-wallet
        scryptsy